+++ dwre_tools/bmtools.py
@@ -6,6 +6,7 @@ import json
 import yaml
 import re
 import zipfile
+import requests
 import io
 from urllib.parse import quote, urlencode, urlparse, parse_qs
 
@@ -16,32 +17,130 @@ CSRF_FINDER2 = re.compile(r"csrf_token=(.*?)\"", re.MULTILINE)
 OPENID_CONNECT_ENDPOINT = "/on/demandware.servlet/dw/oidc/openid_connect_login"
 ACCOUNT_MANAGER_JSON_AUTHENTICATE = "https://account.demandware.com/dwsso/json/realms/root/authenticate"
 ACCOUNT_MANAGER_COOKIE_NAME = "dwAccountManager"
+BM_SESSION_CACHE = dict()
 
+def update_hotfix_path(env, path):
+    use_ocapi = False
+    if "clientID" in env:
+        use_ocapi = True
 
-def get_current_versions(env, session):
-    versions_url = ("https://{}/on/demandware.store/Sites-Site/default/DWREMigrate-Versions"
-                    .format(env["server"]))
-    response = session.get(versions_url)
-
-    if "application/json" not in response.headers['content-type']:
-        raise NotInstalledException("Server response is not json;" +
-                                    "DWRE Tools is likely not installed")
+    if use_ocapi:
+        instance_type = "development"
+        if "instanceType" in env:
+            instance_type = env["instanceType"]
+        session = requests.session()
+        authenticate_session_from_env(env, session)
+        resp = session.patch("https://{}/s/-/dw/data/v20_8/global_preferences/preference_groups/dwreMigrate/{}".format(env["server"], instance_type), json={
+            "c_dwreMigrateHotfixes": path
+        })
+        resp.raise_for_status()
     else:
-        tool_version = response.json()["toolVersion"]
-        migration_version = response.json()["migrationVersion"]
-        cartridge_version = response.json().get('cartridgeVersion')
-        current_migration_path = None
-        hotfixes = []
-        if "migrationPath" in response.json() and response.json()["migrationPath"]:
-            current_migration_path = response.json()["migrationPath"].split(',')
-        if "dwreMigrateHotfixes" in response.json() and response.json()["dwreMigrateHotfixes"]:
-            hotfixes = response.json()["dwreMigrateHotfixes"].split(',')
-        return (tool_version, migration_version, current_migration_path,
-                cartridge_version, hotfixes)
+        bmsession = login_business_manager(env)
+        response = bmsession.post("https://{}/on/demandware.store/Sites-Site/default/DWREMigrate-UpdateVersion".format(env["server"]),
+                                  data={"dwreMigrateHotfixes" : path})
+        response.raise_for_status()
 
 
+def update_current_version(env, id, path):
+    use_ocapi = False
+    if "clientID" in env:
+        use_ocapi = True
 
-def login_via_account_manager(env, session):
+    if use_ocapi:
+        instance_type = "development"
+        if "instanceType" in env:
+            instance_type = env["instanceType"]
+        session = requests.session()
+        authenticate_session_from_env(env, session)
+        resp = session.patch("https://{}/s/-/dw/data/v20_8/global_preferences/preference_groups/dwreMigrate/{}".format(env["server"], instance_type), json={
+            "c_dwreMigrateCurrentVersion": id,
+            "c_dwreMigrateVersionPath": path
+        })
+        resp.raise_for_status()
+    else:
+        bmsession = login_business_manager(env)
+        response = bmsession.post("https://{}/on/demandware.store/Sites-Site/default/DWREMigrate-UpdateVersion".format(env["server"]),
+                                  data={"NewVersion" : id, "NewVersionPath" : path})
+        response.raise_for_status()
+
+
+def get_current_versions(env):
+    use_ocapi = False
+    if "clientID" in env:
+        use_ocapi = True
+
+    if use_ocapi:
+        instance_type = "development"
+        if "instanceType" in env:
+            instance_type = env["instanceType"]
+        session = requests.session()
+        authenticate_session_from_env(env, session)
+        resp = session.get("https://{}/s/-/dw/data/v20_8/global_preferences/preference_groups/dwreMigrate/{}".format(env["server"], instance_type))
+        resp.raise_for_status()
+        j = resp.json()
+        return (
+            j.get("c_dwreMigrateToolVersion"),
+            j.get("c_dwreMigrateCurrentVersion"),
+            j.get("c_dwreMigrateVersionPath").split(","),
+            "2",
+            j.get("c_dwreMigrateHotfixes").split(',')
+        )
+    else:
+        bmsession = login_business_manager(env)
+
+        versions_url = ("https://{}/on/demandware.store/Sites-Site/default/DWREMigrate-Versions"
+                        .format(env["server"]))
+        response = bmsession.get(versions_url)
+
+        if "application/json" not in response.headers['content-type']:
+            raise NotInstalledException("Server response is not json;" +
+                                        "DWRE Tools is likely not installed")
+        else:
+            tool_version = response.json()["toolVersion"]
+            migration_version = response.json()["migrationVersion"]
+            cartridge_version = response.json().get('cartridgeVersion')
+            current_migration_path = None
+            hotfixes = []
+            if "migrationPath" in response.json() and response.json()["migrationPath"]:
+                current_migration_path = response.json()["migrationPath"].split(',')
+            if "dwreMigrateHotfixes" in response.json() and response.json()["dwreMigrateHotfixes"]:
+                hotfixes = response.json()["dwreMigrateHotfixes"].split(',')
+            return (tool_version, migration_version, current_migration_path,
+                    cartridge_version, hotfixes)
+
+
+# Authenticate a session using account manager client credentials
+def authenticate_session_from_env(env, session):
+    auth = (env["clientID"], env["clientPassword"])
+    resp = requests.post("https://account.demandware.com/dwsso/oauth2/access_token", data={"grant_type":"client_credentials"}, auth=auth)
+    resp.raise_for_status()
+    j = resp.json()
+    access_token = j.get("access_token")
+    expiration_seconds = j.get("expires_in");
+    session.headers.update({"Authorization": "Bearer " + access_token})
+    session.headers.update({"x-dw-client-id": env["clientID"]})
+
+
+def authenticate_webdav_session(env):
+    session = requests.session()
+    session.verify = env["verify"]
+    session.cert = env["cert"]
+
+    use_ocapi = False
+    if "clientID" in env:
+        use_ocapi = True
+
+    if use_ocapi:
+        authenticate_session_from_env(env, session)
+    else:
+        session.auth = (env["username"], env["password"])
+    return session
+
+
+def login_via_account_manager(env):
+    session = requests.session()
+    session.verify = env["verify"]
+    session.cert = env["cert"]
     data = dict(
         source='LOGIN_DEFAULT',
         supportLogin='true',
@@ -194,11 +293,21 @@ def login_via_account_manager(env, session):
     else:
         print(resp.text)
         raise RuntimeError("Can't find CSRF")
+    return session
 
 
-def login_business_manager(env, session):
+def login_business_manager(env):
+    global BM_SESSION_CACHE
+    session = requests.session()
+    session.verify = env["verify"]
+    session.cert = env["cert"]
+    if env["server"] in BM_SESSION_CACHE:
+        return BM_SESSION_CACHE[env["server"]]
+
     if env.get('useAccountManager'):
-        return login_via_account_manager(env, session)
+        s = login_via_account_manager(env)
+        BM_SESSION_CACHE[env["server"]] = s
+        return s
 
     response = session.post("https://{}/on/demandware.store/Sites-Site/default/ViewApplication-ProcessLogin".format(env["server"]),
                             data=dict(
@@ -210,7 +319,9 @@ def login_business_manager(env, session):
     response.raise_for_status()
 
     if 'redirectform' in response.text:
-        return login_via_account_manager(env, session)
+        s = login_via_account_manager(env)
+        BM_SESSION_CACHE[env["server"]] = s
+        return s
 
     if "Please create a new password." in response.text:
         raise RuntimeError("Password Expired; New password required")
@@ -248,6 +359,10 @@ def login_business_manager(env, session):
         print(response.text)
         raise RuntimeError("Can't find CSRF")
 
+    BM_SESSION_CACHE[env["server"]] = session
+    return session
+
+
 def select_site(env, session, site_uuid):
     data = {
         "SelectedSiteID": site_uuid
@@ -257,17 +372,72 @@ def select_site(env, session, site_uuid):
     resp.raise_for_status()
 
 
-def activate_code_version(env, session, code_version):
-    session.post("https://{}/on/demandware.store/Sites-Site/default/ViewCodeDeployment-Activate"
-                 .format(env["server"]), data=dict(
-                     CodeVersionID=code_version))
+def activate_code_version(env, code_version):
+    use_ocapi = False
+    if "clientID" in env:
+        use_ocapi = True
+
+    if use_ocapi:
+        session = requests.session()
+        authenticate_session_from_env(env, session)
+        resp = session.patch("https://{}/s/-/dw/data/v20_8/code_versions/{}".format(env["server"], code_version), json={
+            "active": True
+        })
+        resp.raise_for_status()
+        print(resp.json())
+    else:
+        bmsession = login_business_manager(env)
+        bmsession.post("https://{}/on/demandware.store/Sites-Site/default/ViewCodeDeployment-Activate"
+                       .format(env["server"]), data=dict(
+                           CodeVersionID=code_version))
+
+
+def import_site_package(env, filename):
+    use_ocapi = False
+    if "clientID" in env:
+        use_ocapi = True
+
+    if use_ocapi:
+        session = requests.session()
+        authenticate_session_from_env(env, session)
+        resp = session.post("https://{}/s/-/dw/data/v20_8/jobs/sfcc-site-archive-import/executions".format(env["server"]), json={
+            "file_name": filename + ".zip"
+        })
+        resp.raise_for_status()
+        j = resp.json()
+        job_id = j.get("id")
+        finished = False
+
+        # try our best to find the link
+        while not finished:
+            time.sleep(2)
+            resp = session.get("https://{}/s/-/dw/data/v20_8/jobs/sfcc-site-archive-import/executions/{}".format(env["server"], job_id))
+            resp.raise_for_status()
+            j = resp.json()
+            exec_status = j.get("execution_status")
+            log_file_name = j.get("log_file_name")
+            status = j.get("status")
+            if exec_status == "aborted" or status == "ERROR":
+                error = "Unknown"
+                if len(j.get("step_executions")) > 0:
+                    error = j.get("step_executions")[0].get("exit_status").get("message")
+                raise RuntimeError("Failure to import %s. Check job execution log %s.\n%s" % (filename, log_file_name, error))
+            elif exec_status == "finished":
+                finished = True
+    else:
+        bmsession = login_business_manager(env)
+
+        response = bmsession.post(
+            "https://{}/on/demandware.store/Sites-Site/default/ViewSiteImpex-Dispatch".format(env["server"]),
+            data={"import": "", "ImportFileName": filename + ".zip", "realmUse": "False"})
+        response.raise_for_status()
+
+        wait_for_import(env, bmsession, filename)
 
 
 def wait_for_import(env, session, filename):
     log_link = None
     retries = 0
-
-    # TODO detect import in progress
 
     # try our best to find the link
     while not log_link and retries < 600:
@@ -310,7 +480,8 @@ def get_list_data_units(env, session):
     return data_units
 
 
-def export_data_units(env, session, units, filename):
+def export_data_units(env, units, filename):
+    session = login_business_manager(env)
     data = [ ('SelectedExportUnit', u) for u in units ]
     data.append( ('export', '') )
     data.append( ('exportFile', filename) )
@@ -320,9 +491,11 @@ def export_data_units(env, session, units, filename):
     resp.raise_for_status()
 
 
-def wait_for_export(env, session, filename):
+def wait_for_export(env, filename):
     log_link = None
     retries = 0
+
+    session = login_business_manager(env)
 
     # try our best to find the link
     while not log_link and retries < 1000:
@@ -344,9 +517,9 @@ def wait_for_export(env, session, filename):
             time.sleep(2)
 
 
-def get_export_zip(env, bmsession, webdavsession, export_units, filename):
-    export_data_units(env, bmsession, export_units, filename)
-    wait_for_export(env, bmsession, filename)
+def get_export_zip(env, webdavsession, export_units, filename):
+    export_data_units(env, export_units, filename)
+    wait_for_export(env, filename)
 
     dest_url = ("https://{0}/on/demandware.servlet/webdav/Sites/Impex/src/instance/{1}.zip"
                 .format(env["server"], filename))
@@ -357,3 +530,16 @@ def get_export_zip(env, bmsession, webdavsession, export_units, filename):
     resp = webdavsession.delete(dest_url)
     resp.raise_for_status()
     return zip_file
+
+
+def get_install_export_zip(env, webdavsession, filename):
+    if "clientID" in env:
+        use_ocapi = True
+    if use_ocapi:
+        # TODO
+        pass
+    else:
+        zip_file = get_export_zip(env, webdavsession,
+                                  export_units=["AccessRoleExport", "GlobalPreferencesExport"],
+                                  filename=filename)
+        return zip_file