+++ dwre_tools/tail.py
@@ -10,6 +10,7 @@ import functools
 from colorama import Fore, Back, Style
 
 from .webdav import get_directories, get_files
+from .bmtools import authenticate_webdav_session
 
 def compare_file_times(f1, f2):
     if f1[1] > f2[1]:
@@ -19,9 +20,8 @@ def compare_file_times(f1, f2):
     return 0
 
 
-def latest_logs(server, username, password, filters, verify, cert):
-    response = requests.request('PROPFIND', 'https://' + server + '/on/demandware.servlet/webdav/Sites/Logs/',
-                                auth=(username, password), verify=verify, cert=cert)
+def latest_logs(session, server, filters):
+    response = session.request('PROPFIND', 'https://' + server + '/on/demandware.servlet/webdav/Sites/Logs/')
     response.raise_for_status()
     x = ET.fromstring(response.content)
     log_files = get_files(x)
@@ -65,16 +65,13 @@ def tail_log_file(name, content):
 def tail_logs(env, filters, interval):
     # TODO allow usage with client cert, noverify
     server = env["server"]
-    username = env["username"]
-    password = env["password"]
-    verify = env["verify"]
-    cert = env["cert"]
+    session = authenticate_webdav_session(env)
 
-    log_files = latest_logs(server, username, password, filters, verify, cert)
+    log_files = latest_logs(session, server, filters)
     urls = ['https://' + server + '/on/demandware.servlet/webdav/Sites/Logs/' + log[0] for
             log in log_files]
 
-    initial = [requests.head(url, auth=(username, password), verify=verify, cert=cert) for url in urls]
+    initial = [session.head(url) for url in urls]
     [r.raise_for_status() for r in initial]
 
     lengths = [i.headers.get("Content-Length") for i in initial]
@@ -82,7 +79,7 @@ def tail_logs(env, filters, interval):
 
     # get initial for last line purposes (for some reason this returns diff content lengths so we
     # can't use it for the initial length calc
-    tail_requests = [requests.get(url, auth=(username, password), verify=verify, cert=cert) for url in urls]
+    tail_requests = [session.get(url) for url in urls]
     [tail_log_file(log[0], r.content.decode('utf8', errors='replace')) for r, log in zip(tail_requests, log_files)]
 
     try:
@@ -91,9 +88,9 @@ def tail_logs(env, filters, interval):
 
             for i, log in enumerate(log_files):
                 url = 'https://' + server + '/on/demandware.servlet/webdav/Sites/Logs/' + log[0]
-                response = requests.get(url, auth=(username, password), headers={
+                response = session.get(url, headers={
                     "range": "bytes=%s-" % (lengths[i])
-                }, verify=verify, cert=cert)
+                })
                 if response.status_code == 416:
                     continue
                 response.raise_for_status()